+++ docs/ChangeLog.md
@@ -54,6 +54,7 @@
 - highlight: reflect the top-level positional parameters `#D2246` f08e8f08
 - color: adjust default fg values in faces and add `bleopt color_scheme` (requested by mattmc3) `#D2248` e4cce0ea 5f5554a8
 - highlight: add `bleopt highlight_eval_word_limit` (motivated by orionalves) `#D2256` 6833bdf8
+- progcomp: support `complete -E` `#D2257` xxxxxxxx
 
 ## Changes